refactor existing test instead
