+++ data/beaconrestapi/src/integration-test/resources/tech/pegasys/teku/beaconrestapi/beacon/paths/_eth_v2_beacon_blocks.json
@@ -83,7 +83,7 @@
         }
       },
       "503" : {
-        "description" : "Beacon node is currently syncing.",
+        "description" : "Beacon node is currently syncing and not serving requests",
         "content" : {
           "application/json" : {
             "schema" : {
